requirements.txt
fastapi
uvicorn
gunicorn
pandas
langchain==0.1.17rc1
langchain-openai==0.1.6
langchain-groq==0.1.3
langchain-anthropic==0.1.11
python-multipart
starlette
sqlalchemy
asyncpg
greenlet
openpyxl
psycopg2
azure-search-documents
azure-identity
numpy
azure-storage-blob
PyPDF2
openai
flower
celery[redis]
pymongo
python-docx
strictjson